Moves errors to the models package (#191)
