+++ composeApp/src/commonMain/kotlin/com/revenuecat/purchases/kmp/sample/rememberLoggingPaywallListener.kt
@@ -2,9 +2,9 @@ package com.revenuecat.purchases.kmp.sample
 
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.remember
-import com.revenuecat.purchases.kmp.models.Package
-import com.revenuecat.purchases.kmp.PurchasesError
 import com.revenuecat.purchases.kmp.models.CustomerInfo
+import com.revenuecat.purchases.kmp.models.Package
+import com.revenuecat.purchases.kmp.models.PurchasesError
 import com.revenuecat.purchases.kmp.models.StoreTransaction
 import com.revenuecat.purchases.kmp.ui.revenuecatui.PaywallListener
