+++ either/api/either.api
@@ -20,7 +20,7 @@ public final class com/revenuecat/purchases/kmp/either/EitherKt {
 }
 
 public final class com/revenuecat/purchases/kmp/either/FailedPurchase {
-	public final fun getError ()Lcom/revenuecat/purchases/kmp/PurchasesError;
+	public final fun getError ()Lcom/revenuecat/purchases/kmp/models/PurchasesError;
 	public final fun getUserCancelled ()Z
 }
