+++ mappings/src/androidMain/kotlin/com/revenuecat/purchases/kmp/mappings/errors.android.kt
@@ -1,7 +1,7 @@
 package com.revenuecat.purchases.kmp.mappings
 
-import com.revenuecat.purchases.kmp.PurchasesError
-import com.revenuecat.purchases.kmp.PurchasesErrorCode
+import com.revenuecat.purchases.kmp.models.PurchasesError
+import com.revenuecat.purchases.kmp.models.PurchasesErrorCode
 import com.revenuecat.purchases.PurchasesError as RcPurchasesError
 import com.revenuecat.purchases.PurchasesErrorCode as RcPurchasesErrorCode
