+++ models/src/commonMain/kotlin/com/revenuecat/purchases/kmp/PurchasesDelegate.kt
@@ -1,6 +1,7 @@
 package com.revenuecat.purchases.kmp
 
 import com.revenuecat.purchases.kmp.models.CustomerInfo
+import com.revenuecat.purchases.kmp.models.PurchasesError
 import com.revenuecat.purchases.kmp.models.StoreProduct
 import com.revenuecat.purchases.kmp.models.StoreTransaction
