+++ models/src/commonMain/kotlin/com/revenuecat/purchases/kmp/models/PurchasesException.kt
@@ -1,4 +1,4 @@
-package com.revenuecat.purchases.kmp
+package com.revenuecat.purchases.kmp.models
 
 /**
  * An exception wrapping a [PurchasesError].
