+++ revenuecatui/api/revenuecatui.klib.api
@@ -9,10 +9,10 @@
 abstract interface com.revenuecat.purchases.kmp.ui.revenuecatui/PaywallListener { // com.revenuecat.purchases.kmp.ui.revenuecatui/PaywallListener|null[0]
     open fun onPurchaseCancelled() // com.revenuecat.purchases.kmp.ui.revenuecatui/PaywallListener.onPurchaseCancelled|onPurchaseCancelled(){}[0]
     open fun onPurchaseCompleted(com.revenuecat.purchases.kmp.models/CustomerInfo, com.revenuecat.purchases.kmp.models/StoreTransaction) // com.revenuecat.purchases.kmp.ui.revenuecatui/PaywallListener.onPurchaseCompleted|onPurchaseCompleted(com.revenuecat.purchases.kmp.models.CustomerInfo;com.revenuecat.purchases.kmp.models.StoreTransaction){}[0]
-    open fun onPurchaseError(com.revenuecat.purchases.kmp/PurchasesError) // com.revenuecat.purchases.kmp.ui.revenuecatui/PaywallListener.onPurchaseError|onPurchaseError(com.revenuecat.purchases.kmp.PurchasesError){}[0]
+    open fun onPurchaseError(com.revenuecat.purchases.kmp.models/PurchasesError) // com.revenuecat.purchases.kmp.ui.revenuecatui/PaywallListener.onPurchaseError|onPurchaseError(com.revenuecat.purchases.kmp.models.PurchasesError){}[0]
     open fun onPurchaseStarted(com.revenuecat.purchases.kmp.models/Package) // com.revenuecat.purchases.kmp.ui.revenuecatui/PaywallListener.onPurchaseStarted|onPurchaseStarted(com.revenuecat.purchases.kmp.models.Package){}[0]
     open fun onRestoreCompleted(com.revenuecat.purchases.kmp.models/CustomerInfo) // com.revenuecat.purchases.kmp.ui.revenuecatui/PaywallListener.onRestoreCompleted|onRestoreCompleted(com.revenuecat.purchases.kmp.models.CustomerInfo){}[0]
-    open fun onRestoreError(com.revenuecat.purchases.kmp/PurchasesError) // com.revenuecat.purchases.kmp.ui.revenuecatui/PaywallListener.onRestoreError|onRestoreError(com.revenuecat.purchases.kmp.PurchasesError){}[0]
+    open fun onRestoreError(com.revenuecat.purchases.kmp.models/PurchasesError) // com.revenuecat.purchases.kmp.ui.revenuecatui/PaywallListener.onRestoreError|onRestoreError(com.revenuecat.purchases.kmp.models.PurchasesError){}[0]
     open fun onRestoreStarted() // com.revenuecat.purchases.kmp.ui.revenuecatui/PaywallListener.onRestoreStarted|onRestoreStarted(){}[0]
 }
 final class com.revenuecat.purchases.kmp.ui.revenuecatui/PaywallOptions { // com.revenuecat.purchases.kmp.ui.revenuecatui/PaywallOptions|null[0]
